Rename license files to make this package's docs visible on pkg.go.dev
The pkg.go.dev site will not display a package's docs unless it can determine
that package's license. It's ability to do that is based in part on the names
of the license files. See https://pkg.go.dev/license-policy for details.

This commit renames this repo's license files to names that pkg.go.dev should
recognize.

Signed-off-by: Dave Rolsky <daver@activestate.com>
Signed-off-by: Ian Ling <ian@iancaling.com>